requirements.txt
requests>=2.32.5
PyYAML>=6.0.3
